requirements.txt
langchain>=0.1.0
langserve>=0.0.30
langchain-openai>=0.0.5
yfinance>=0.2.31
pandas>=2.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
langchain-community==0.0.13
